requirements.txt
fastapi
uvicorn
python-dotenv
google-generativeai
pandas
numpy
scipy
matplotlib
seaborn
beautifulsoup4
requests
lxml
duckdb
Pillow
scikit-learn
statsmodels
langchain-community
sentence-transformers
faiss-cpu
torch
pytesseract
langchain-google-genai